Hooray!!!
First functioning, runnable TrackBuilder, Track and TrackBuilderTest class.

Edited height_and_width_test.txt and added track_fill_test.txt test tracks.

Last Junit test (fillMapTest) fails, but not because the data doesn't match - it's because the pointer name (or something, idunno doesn't match. Try to resolve this later but for now i'm really happy I got this beast to work.

Still need to do TrackTest class, but since it's so closely coupled to TrackBuilder (and in this one instance it's actually a GOOD thing) it's just mere formality.